[Misc] fake commit for testing release builds

+++ src/apps/depth2x/converter_curvature.cpp
@@ -15,7 +15,7 @@ bool gauss_curv_converter::process_file(cv::Mat depth_image, int idx) const
     const cv::Mat gauss =
         depth_to_gaussian_curvature<double, double>(depth_image, intrinsic);
     cv::Mat gauss_conv;
-    gauss.convertTo(gauss_conv, CV_8S);
+    gauss.convertTo(gauss_conv, CV_16S);
     bool success = cv::imwrite(fmt::format(_files.output, idx), gauss_conv);
 
     return success;
@@ -30,7 +30,7 @@ bool mean_curv_converter::process_file(cv::Mat depth_image, int idx) const
     const cv::Mat mean =
         depth_to_mean_curvature<double, double>(depth_image, intrinsic);
     cv::Mat mean_conv;
-    mean.convertTo(mean_conv, CV_8S);
+    mean.convertTo(mean_conv, CV_16S);
     bool success = cv::imwrite(fmt::format(_files.output, idx), mean_conv);
 
     return success;
